# api/swagger.yaml
openapi: 3.0.1
info:
  title: Product Catalog Management
  description: "## TMF API Reference: TMF620 - Product Catalog Management\n\n### Release\
    \ : 19.0 - June 2019\n\nProduct Catalog API is one of Catalog Management API Family.\
    \ Product Catalog API goal is to provide a catalog of products. \n\n### Operations\n\
    Product Catalog API performs the following operations on the resources :\n- Retrieve\
    \ an entity or a collection of entities depending on filter criteria\n- Partial\
    \ update of an entity (including updating rules)\n- Create an entity (including\
    \ default values and creation rules)\n- Delete an entity\n- Manage notification\
    \ of events"
  version: 4.0.0
servers:
- url: https://serverRoot/tmf-api/productCatalogManagement/v4/
...
paths:
...
  /importJob:
    get:
      tags:
      - importJob
      summary: List or find ImportJob objects
      description: This operation list or find ImportJob entities
      operationId: listImportJob
      parameters:
      - name: fields
        in: query
        description: Comma-separated properties to be provided in response
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: offset
        in: query
        description: Requested index for start of resources to be provided in response
        required: false
        style: form
        explode: true
        schema:
          type: integer
      - name: limit
        in: query
        description: Requested number of resources to be provided in response
        required: false
        style: form
        explode: true
        schema:
          type: integer
      responses:
        200:
          description: Success
          headers:
            X-Total-Count:
              description: Total number of items matching criteria
              style: simple
              explode: false
              schema:
                type: integer
            X-Result-Count:
              description: Actual number of items returned in the response body
              style: simple
              explode: false
              schema:
                type: integer
          content:
            application/json;charset=utf-8:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ImportJob'
                x-content-type: application/json;charset=utf-8
        400:
          description: Bad Request
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        401:
          description: Unauthorized
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        403:
          description: Forbidden
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        404:
          description: Not Found
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        405:
          description: Method Not allowed
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        409:
          description: Conflict
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Internal Server Error
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
      x-swagger-router-controller: ImportJob
    post:
      tags:
      - importJob
      summary: Creates a ImportJob
      description: This operation creates a ImportJob entity.
      operationId: createImportJob
      requestBody:
        description: The ImportJob to be created
        content:
          application/json;charset=utf-8:
            schema:
              $ref: '#/components/schemas/ImportJob_Create'
        required: true
      responses:
        201:
          description: Created
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/ImportJob'
        400:
          description: Bad Request
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        401:
          description: Unauthorized
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        403:
          description: Forbidden
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        405:
          description: Method Not allowed
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        409:
          description: Conflict
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Internal Server Error
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
      x-codegen-request-body-name: importJob
      x-swagger-router-controller: ImportJob
  /importJob/{id}:
    get:
      tags:
      - importJob
      summary: Retrieves a ImportJob by ID
      description: This operation retrieves a ImportJob entity. Attribute selection
        is enabled for all first level attributes.
      operationId: retrieveImportJob
      parameters:
      - name: id
        in: path
        description: Identifier of the ImportJob
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: fields
        in: query
        description: Comma-separated properties to provide in response
        required: false
        style: form
        explode: true
        schema:
          type: string
      responses:
        200:
          description: Success
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/ImportJob'
        400:
          description: Bad Request
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        401:
          description: Unauthorized
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        403:
          description: Forbidden
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        404:
          description: Not Found
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        405:
          description: Method Not allowed
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        409:
          description: Conflict
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Internal Server Error
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
      x-swagger-router-controller: ImportJob
    delete:
      tags:
      - importJob
      summary: Deletes a ImportJob
      description: This operation deletes a ImportJob entity.
      operationId: deleteImportJob
      parameters:
      - name: id
        in: path
        description: Identifier of the ImportJob
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        204:
          description: Deleted
          content: {}
        400:
          description: Bad Request
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        401:
          description: Unauthorized
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        403:
          description: Forbidden
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        404:
          description: Not Found
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        405:
          description: Method Not allowed
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        409:
          description: Conflict
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Internal Server Error
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
      x-swagger-router-controller: ImportJob
...
components:
  schemas:
...
    ImportJob:
      type: object
      properties:
        id:
          type: string
          description: Identifier of the import job
        href:
          type: string
          description: Reference of the import job
        completionDate:
          type: string
          description: Date at which the job was completed
          format: date-time
        contentType:
          type: string
          description: Indicates the format of the imported data
        creationDate:
          type: string
          description: Date at which the job was created
          format: date-time
        errorLog:
          type: string
          description: Reason for failure if status is failed
        path:
          type: string
          description: URL of the root resource where the content of the file specified
            by the import job must be applied
        url:
          type: string
          description: URL of the file containing the data to be imported
        status:
          $ref: '#/components/schemas/JobStateType'
        '@baseType':
          type: string
          description: When sub-classing, this defines the super-class
        '@schemaLocation':
          type: string
          description: A URI to a JSON-Schema file that defines additional attributes
            and relationships
          format: uri
        '@type':
          type: string
          description: When sub-classing, this defines the sub-class entity name
      description: Represents a task used to import resources from a file
      example:
        path: path
        '@baseType': '@baseType'
        '@type': '@type'
        completionDate: 2000-01-23T04:56:07.000+00:00
        errorLog: errorLog
        id: id
        href: href
        creationDate: 2000-01-23T04:56:07.000+00:00
        '@schemaLocation': http://example.com/aeiou
        contentType: contentType
        url: url
        status: Not Started
    ImportJob_Create:
      required:
      - url
      type: object
      properties:
        completionDate:
          type: string
          description: Date at which the job was completed
          format: date-time
        contentType:
          type: string
          description: Indicates the format of the imported data
        creationDate:
          type: string
          description: Date at which the job was created
          format: date-time
        errorLog:
          type: string
          description: Reason for failure if status is failed
        path:
          type: string
          description: URL of the root resource where the content of the file specified
            by the import job must be applied
        url:
          type: string
          description: URL of the file containing the data to be imported
        status:
          $ref: '#/components/schemas/JobStateType'
        '@baseType':
          type: string
          description: When sub-classing, this defines the super-class
        '@schemaLocation':
          type: string
          description: A URI to a JSON-Schema file that defines additional attributes
            and relationships
          format: uri
        '@type':
          type: string
          description: When sub-classing, this defines the sub-class entity name
      description: |-
        Represents a task used to import resources from a file
        Skipped properties: id,href
    JobStateType:
      type: string
      description: Valid values for the state of a batch job (e.g. catalog import)
      enum:
      - Not Started
      - Running
      - Succeeded
      - Failed
...
    Error:
      required:
      - code
      - reason
      type: object
      properties:
        code:
          type: string
          description: Application relevant detail, defined in the API or a common
            list.
        reason:
          type: string
          description: Explanation of the reason for the error which can be shown
            to a client user.
        message:
          type: string
          description: More details and corrective actions related to the error which
            can be shown to a client user.
        status:
          type: string
          description: HTTP Error code extension
        referenceError:
          type: string
          description: URI of documentation describing the error.
          format: uri
        '@baseType':
          type: string
          description: When sub-classing, this defines the super-class.
        '@schemaLocation':
          type: string
          description: A URI to a JSON-Schema file that defines additional attributes
            and relationships
          format: uri
        '@type':
          type: string
          description: When sub-classing, this defines the sub-class entity name.
      description: Used when an API throws an Error, typically with a HTTP error response-code
        (3xx, 4xx, 5xx)